+++ http/src/main/resources/reference.conf
@@ -166,17 +166,6 @@ azure {
      disk-controlled-resource-desc = "Azure Disk"
      vm-controlled-resource-desc = "Azure Vm"
    }
-
-   wsm {
-      uri = "https://localhost:8000"
-   }
-
-  # We need the leo azure entity for this
-   app-registration {
-     client-id = ""
-     client-secret = ""
-     managed-app-tenant-id = ""
-   }
 }
 
 dateAccessedUpdater {
