+++ http/src/test/resources/reference.conf
@@ -293,21 +293,9 @@ akka.ssl-config {
 azure {
    monitor {
     poll-status {
-      max-attempts = 120 # 15 seconds * 120 is 30 min
+      max-attempts = 120
       interval = 1 seconds
     }
-   }
-
-   runtime-defaults {
-     ip-name-prefix = "ip"
-     ip-controlled-resource-desc = "Azure Ip"
-     network-controlled-resource-desc = "Azure Network"
-     network-name-prefix = "network"
-     subnet-name-prefix = "subnet"
-     address-space-cidr = "192.168.0.0/16"
-     subnet-address-cidr = "192.168.0.0/24"
-     disk-controlled-resource-desc = "Azure Disk"
-     vm-controlled-resource-desc = "Azure Vm"
    }
 
    wsm {
